backend/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.0.1
six==1.16.0
Werkzeug==2.0.3
matplotlib
torch
torchvision 
torchaudio
requests